requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
redis>=5.0.1
numpy>=1.26.0
pydantic>=2.10.0
requests>=2.31.0
python-dotenv>=1.0.0

# RedisVL for semantic routing
redisvl>=0.3.0
sentence-transformers>=2.3.1

# Transformers for language detection
transformers>=4.36.0
torch>=2.1.0

# OpenAI for chat route + concierge tool-calling
openai>=1.10.0
langchain-core>=0.3.78,<0.4
langchain-openai>=0.2.0,<0.4
langchain-redis>=0.2.0,<0.4

